requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
cryptography==41.0.4
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Cors==4.0.0
flask-sqlalchemy==3.0.5
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.4
openai==1.3.7
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.0.3
Pillow==10.0.0
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.2
pydantic-core==2.14.5
PyJWT==2.8.0
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
regex==2023.12.25
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.19
tiktoken==0.5.2
tqdm==4.66.1
treelib==1.6.4
trio==0.23.2
trio-websocket==0.11.1
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.1.0
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.4
zipp==3.16.2
